+++ CHANGELOG.md
@@ -7,11 +7,21 @@ This project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+## [2.4.0] - 2024-03-20
+
+### Added
+
+- Added support for uploading with user token and app ID.
+
+### Changed
+
+- Modified help and error messages to more closely align with Waldo CLI.
+
 ## [2.3.1] - 2024-03-15
 
 ### Changed
 
-- Fix encoding of upload errors.
+- Fixed encoding of upload errors.
 
 ## [2.3.0] - 2024-02-26
 
@@ -71,7 +81,8 @@ This project adheres to [Semantic Versioning].
 
 Initial public release of rewritten agent.
 
-[Unreleased]:   https://github.com/waldoapp/waldo-go-agent/compare/2.3.1...HEAD
+[Unreleased]:   https://github.com/waldoapp/waldo-go-agent/compare/2.4.0...HEAD
+[2.4.0]:        https://github.com/waldoapp/waldo-go-agent/compare/2.3.1...2.4.0
 [2.3.1]:        https://github.com/waldoapp/waldo-go-agent/compare/2.3.0...2.3.1
 [2.3.0]:        https://github.com/waldoapp/waldo-go-agent/compare/2.2.0...2.3.0
 [2.2.0]:        https://github.com/waldoapp/waldo-go-agent/compare/2.1.3...2.2.0